chatbot/requirements.txt
# LangChain Core
langchain
langchain-core
langchain_community
streamlit
# Hugging Face Integration
langchain-huggingface
huggingface-hub

# Environment Variable Management
python-dotenv
